@@ -3,36 +3,40 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@emailjs/browser": "^3.10.0",
-    "@fortawesome/fontawesome-svg-core": "^6.1.1",
-    "@fortawesome/free-solid-svg-icons": "^6.1.1",
-    "@fortawesome/react-fontawesome": "^0.1.18",
-    "@react-pdf/renderer": "^3.1.12",
-    "@reduxjs/toolkit": "^1.9.2",
-    "@testing-library/jest-dom": "^5.16.2",
-    "@testing-library/react": "^12.1.4",
-    "@testing-library/user-event": "^13.5.0",
+    "@emailjs/browser": "^4.3.3",
+    "@fortawesome/fontawesome-svg-core": "^6.5.2",
+    "@fortawesome/free-solid-svg-icons": "^6.5.2",
+    "@fortawesome/react-fontawesome": "^0.2.0",
+    "@react-pdf/renderer": "^3.4.4",
+    "@react-three/drei": "^9.105.6",
+    "@react-three/fiber": "^8.16.6",
+    "@reduxjs/toolkit": "^2.2.4",
+    "@testing-library/jest-dom": "^6.4.5",
+    "@testing-library/react": "^15.0.7",
+    "@testing-library/user-event": "^14.5.2",
     "aos": "^3.0.0-beta.6",
-    "axios": "^1.4.0",
-    "classnames": "^2.3.1",
+    "axios": "^1.6.8",
+    "classnames": "^2.5.1",
     "downloadjs": "^1.4.7",
-    "gh-pages": "^3.2.3",
+    "gh-pages": "^6.1.1",
+    "gltfjsx": "^6.2.16",
     "lodash": "^4.17.21",
-    "nuka-carousel": "^5.0.11",
-    "pretendard": "^1.2.2",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-redux": "^8.0.5",
-    "react-router-dom": "^6.2.2",
-    "react-scripts": "5.0.0",
+    "nuka-carousel": "^8.0.1",
+    "pretendard": "^1.3.9",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "react-redux": "^9.1.2",
+    "react-router-dom": "^6.23.1",
+    "react-scripts": "5.0.1",
     "react-transition-group": "^4.4.5",
-    "recharts": "^2.1.9",
-    "redux": "^4.2.1",
+    "recharts": "^2.12.7",
+    "redux": "^5.0.1",
     "redux-devtools-extension": "^2.13.9",
     "redux-logger": "^3.0.6",
-    "smooth-scrollbar": "^8.7.4",
+    "smooth-scrollbar": "^8.8.4",
+    "three": "^0.164.1",
     "use-interval": "^1.4.0",
-    "web-vitals": "^2.1.4"
+    "web-vitals": "^3.5.2"
   },
   "scripts": {
     "start": "react-scripts start",
